fix: split FAST v2 and v3 plugins
