+++ crates/webui-cli/src/commands/serve.rs
@@ -16,7 +16,8 @@ use std::sync::{Arc, Mutex};
 use std::time::Duration;
 use webui::WebUIHandler;
 use webui_dev_server::{spawn_watcher, sse_handler, LiveReload, WatchConfig};
-use webui_handler::plugin::fast::{FastV2HydrationPlugin, FastV3HydrationPlugin};
+use webui_handler::plugin::fast_v2::FastV2HydrationPlugin;
+use webui_handler::plugin::fast_v3::FastV3HydrationPlugin;
 use webui_handler::plugin::webui::WebUIHydrationPlugin;
 use webui_handler::{RenderOptions, ResponseWriter};
 use webui_protocol::WebUIProtocol;
