+++ crates/webui-handler/benches/handler_bench.rs
@@ -5,7 +5,7 @@ use serde_json::json;
 use serde_json::Value;
 use std::collections::HashMap;
 use std::hint::black_box;
-use webui_handler::plugin::fast::FastV2HydrationPlugin;
+use webui_handler::plugin::fast_v2::FastV2HydrationPlugin;
 use webui_handler::{RenderOptions, ResponseWriter, WebUIHandler};
 use webui_protocol::{
     ComparisonOperator, ConditionExpr, FragmentList, LogicalOperator, WebUIFragment, WebUIProtocol,
